Delete README.md
Team Name: Alpha Techies
Product Name: E-Markety


Project Idea:
The project idea is to develop an innovative web3 platform that tokenizes online marketplaces, enabling the buying and selling of services and goods, with a specific focus on essential items such as food and basic necessities. The primary means of exchange for transactions on this platform will be Hedera Coin, a fast, secure, and low-cost cryptocurrency.

Product Overview:
The goal of our web3 product is to create a decentralized online marketplace that facilitates the buying and selling of services and goods, focusing on essential items such as food and basic necessities. The marketplace will be built on hedera blockchain technology, specifically leveraging Hedera Coin as the primary means of exchange. By tokenizing transactions and incorporating smart contracts, we aim to provide a secure, transparent, and efficient platform for users(buyers and sellers) in the African market to access essential goods and services.

Problem Statement for the Africa Market:
In many parts of Africa, using Nigeria as a case study, accessing essential goods and services can be challenging due to fluctuating high costs, exploitation, lack of reliable online marketplaces and limited infrastructure. 
Traditional markets which seem to be the primary marketplace for Africans tend to be an avenue for increase in the rate of exploitation among the buyers and sellers with irregular skyrocketed high cost.
Also, traditional e-commerce platforms often face difficulties in delivering products to remote areas, and the lack of trust and transparency in transactions can discourage users from engaging in online commerce. 
Furthermore, during the era of no cash flow in Nigeria, a lot of barriers especially financial barriers coupled with limited access to banking services set in making it difficult for people to get basic things which led to an increase in the rate of exploitation, it also hindered the participation of underserved communities in the digital economy.